folder restructure and complete app navigation with authentication validation
