+++ src/App.js
@@ -1,9 +1,8 @@
 import './App.css';
-import Landing from  './Components/Landing';
-import UserAuthentication from  './Components/UserAuthentication';
-import Dashboard from  './Components/Dashboard';
+import Landing from  './Containers/Landing';
+import UserAuthentication from  './Containers/UserAuthentication';
+import Dashboard from  './Containers/Dashboard';
 import React from 'react';
-
 import PrivateRoute from './Routes/PrivateRoute';
 import { Route, Routes } from 'react-router-dom';
 import AuthProvider from './Provider/authProvider';
